refactor cpu tests
